requirements.txt
django
djangorestframework
django-rest-swagger
django-generate-secret-key
pyyaml
gunicorn